requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
Babel==2.9.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
Django==3.1.3
django-braces==1.14.0
django-countries==7.0
django-credit-cards==0.4.1
django-crispy-forms==1.10.0
django-flexible-subscriptions==0.15.1
django-imperavi==0.2.3
django-imperavi-widget==0.0.3
django-libs==2.0.3
django-messages==0.6.0
django-model-utils==4.1.1
-e git://github.com/dokterbob/django-newsletter.git@4a518f5380ed00602fa83b14b1d52842f0845a7a#egg=django_newsletter
django-newsletter-signup==0.6.2
django-newsletter-subscription==0.4.0
django-postman==4.0
django-private-chat==0.3.0
docutils==0.16
filelock==3.0.12
idna==2.10
imagesize==1.2.0
isort==5.6.4
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
ldif3==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
nmap==0.0.1
numpy==1.19.4
packaging==20.8
pandas==1.1.4
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-card-me==0.9.3
python-dateutil==2.8.1
python-gettext==4.0
python-nmap==0.6.1
python3-nmap==1.4.9
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.7.0
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stripe==2.55.0
surlex==0.2.0
toml==0.10.2
unicodecsv==0.14.1
urllib3==1.26.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
websockets==8.1
wrapt==1.12.1